requirements.txt
-i https://pypi.douban.com/simple
aiofiles==0.6.0
aioredis==1.3.1
async-timeout==3.0.1; python_full_version >= '3.5.3'
bcrypt==3.2.0; python_version >= '3.6'
certifi==2020.6.20
cffi==1.14.3
contextvars==2.4; python_version < '3.7'
cryptography==3.2.1; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4'
gunicorn==20.0.4
h11==0.9.0
hiredis==1.1.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
httpcore==0.11.1; python_version >= '3.6'
httptools==0.1.1
httpx==0.15.4; python_version >= '3.6'
idna==2.10; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
immutables==0.14; python_version >= '3.5'
multidict==5.0.0; python_version >= '3.5'
paramiko==2.7.2
pycparser==2.20; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
pyjwt==1.7.1
pymysql==0.10.1
pynacl==1.4.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
rfc3986[idna2008]==1.4.0
sanic-cors==0.10.0.post3
sanic-openapi==0.6.2
sanic-plugins-framework==0.9.4.post1
sanic==20.9.1
six==1.15.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
sniffio==1.2.0; python_version >= '3.5'
sqlalchemy==1.3.20
ujson==4.0.1; sys_platform != 'win32' and implementation_name == 'cpython'
uvloop==0.14.0; sys_platform != 'win32' and implementation_name == 'cpython'
websockets==8.1; python_full_version >= '3.6.1'
